refactors perks
