+++ src/pages/mint-pass/index.tsx
@@ -2,7 +2,7 @@ import Feature from '@/components/BrainPass/Feature'
 import NetworkConnectionInfo from '@/components/Layout/Network/NetworkConnectionInfo'
 import NetworkErrorNotification from '@/components/Layout/Network/NetworkErrorNotification'
 import MintNotification from '@/components/Layout/Nft/MintNotification'
-import { PASS_FEATURES } from '@/data/PassFeatureData'
+import { PASS_FEATURES } from '@/data/PassData'
 import { env } from '@/env.mjs'
 import useBrainPass from '@/hooks/useBrainPass'
 import useNetworkProvider from '@/hooks/useNetworkProvider'
